+++ examples/react/UIShell/src/App.tsx
@@ -16,34 +16,45 @@ import {
   SideNavMenu,
   SideNavMenuItem,
   HeaderDivider,
+  HeaderContainer,
+  HeaderPanel,
+  HeaderPopover,
+  HeaderPopoverActions,
+  HeaderPopoverButton,
+  HeaderPopoverContent,
+  TrialCountdown,
 } from '@carbon-labs/react-ui-shell/es/index';
 import {
-  Content,
-  Grid,
-  Column,
   SkipToContent,
-  HeaderContainer,
   Header,
   HeaderName,
   Theme,
   HeaderMenuButton,
   SideNavDivider,
-  HeaderNavigation,
-  HeaderMenu,
-  HeaderMenuItem,
+  Content,
+  Grid,
+  Column,
   HeaderGlobalBar,
-  ExpandableSearch,
+  HeaderGlobalAction,
+  Switcher,
+  SwitcherItem,
+  SwitcherDivider,
   MenuButton,
   MenuItemRadioGroup,
-  HeaderGlobalAction,
+  ExpandableSearch,
+  Link,
+  Button,
 } from '@carbon/react';
 import {
   Fade,
   SquareOutline,
   Help,
   Notification,
-  IbmWatsonxAssistant,
+  Switcher as SwitcherIcon,
   UserAvatar,
+  Share,
+  User,
+  ShoppingCart,
 } from '@carbon/icons-react';
 
 const StoryContent = () => (
@@ -133,7 +144,12 @@ function App() {
   return (
     <Theme theme="g100">
       <HeaderContainer
-        render={({ isSideNavExpanded, onClickSideNavExpand }) => (
+        render={({
+          isSideNavExpanded,
+          isSwitcherExpanded,
+          onClickSideNavExpand,
+          onClickSwitcherExpand,
+        }) => (
           <>
             <Theme theme="g100">
               <Header aria-label="IBM Platform Name">
@@ -149,16 +165,26 @@ function App() {
                 <HeaderName href="#" prefix="IBM">
                   [Platform]
                 </HeaderName>
-                <HeaderNavigation aria-label="Nav 1">
-                  <HeaderMenuItem href="#">HeaderMenuItem</HeaderMenuItem>
-                  <HeaderMenu aria-label="Link 4" menuLinkName="HeaderMenu">
-                    <HeaderMenuItem href="#">HeaderMenuItem</HeaderMenuItem>
-                    <HeaderMenuItem isActive href="#">
-                      HeaderMenuItem
-                    </HeaderMenuItem>
-                    <HeaderMenuItem href="#">HeaderMenuItem</HeaderMenuItem>
-                  </HeaderMenu>
-                </HeaderNavigation>
+                <HeaderPopover align="bottom">
+                  <HeaderPopoverButton
+                    label="Trial Countdown"
+                    as={Button}
+                    kind="ghost">
+                    <TrialCountdown count={30} />
+                  </HeaderPopoverButton>
+                  <HeaderPopoverContent>
+                    <p>Your trial ends on May 13, 2025</p>
+                    <Link href="#" renderIcon={Share}>
+                      Invite team members
+                    </Link>
+                    <Link href="#" renderIcon={User}>
+                      Contact sales
+                    </Link>
+                    <Button size="sm" renderIcon={ShoppingCart}>
+                      Buy
+                    </Button>
+                  </HeaderPopoverContent>
+                </HeaderPopover>
                 <HeaderGlobalBar>
                   <ExpandableSearch
                     size="lg"
@@ -166,6 +192,43 @@ function App() {
                     closeButtonLabelText="Clear search input"
                     id="search-expandable-1"
                   />
+                  <HeaderGlobalAction
+                    aria-label="Custom action"
+                    tooltipHighContrast={false}>
+                    <SquareOutline size={20} />
+                  </HeaderGlobalAction>
+                  <HeaderPopover align="bottom-right">
+                    <HeaderPopoverButton align="bottom" label="Help">
+                      <Help size={20} />
+                    </HeaderPopoverButton>
+                    <HeaderPopoverContent>
+                      <p>
+                        Lorem ipsum dolor sit amet, di os consectetur adipiscing
+                        elit, sed do eiusmod tempor incididunt ut fsil labore et
+                        dolore magna aliqua.
+                      </p>
+                      <HeaderPopoverActions>
+                        <Link href="#">Link action</Link>
+                        <Button size="sm">Button</Button>
+                      </HeaderPopoverActions>
+                    </HeaderPopoverContent>
+                  </HeaderPopover>
+                  <HeaderPopover align="bottom-right">
+                    <HeaderPopoverButton align="bottom" label="Notifications">
+                      <Notification size={20} />
+                    </HeaderPopoverButton>
+                    <HeaderPopoverContent>
+                      <p>
+                        Lorem ipsum dolor sit amet, di os consectetur adipiscing
+                        elit, sed do eiusmod tempor incididunt ut fsil labore et
+                        dolore magna aliqua.
+                      </p>
+                      <HeaderPopoverActions>
+                        <Link href="#">Link action</Link>
+                        <Button size="sm">Button</Button>
+                      </HeaderPopoverActions>
+                    </HeaderPopoverContent>
+                  </HeaderPopover>
                   <HeaderDivider />
                   <MenuButton
                     menuTarget={headerRef.current}
@@ -194,27 +257,70 @@ function App() {
                     />
                   </MenuButton>
                   <HeaderDivider />
+                  <HeaderPopover align="bottom-right">
+                    <HeaderPopoverButton align="bottom" label="Profile">
+                      <UserAvatar size={20} />
+                    </HeaderPopoverButton>
+                    <HeaderPopoverContent>
+                      <p>
+                        Lorem ipsum dolor sit amet, di os consectetur adipiscing
+                        elit, sed do eiusmod tempor incididunt ut fsil labore et
+                        dolore magna aliqua.
+                      </p>
+                      <HeaderPopoverActions>
+                        <Link href="#">Link action</Link>
+                        <Button size="sm">Button</Button>
+                      </HeaderPopoverActions>
+                    </HeaderPopoverContent>
+                  </HeaderPopover>
+                  {/* <HeaderGlobalAction
+                    aria-label="Custom action"
+                    tooltipHighContrast={false}
+                  >
+                    <SwitcherIcon size={20} />
+                  </HeaderGlobalAction> */}
                   <HeaderGlobalAction
-                    aria-label="Chat"
-                    tooltipHighContrast={false}>
-                    <IbmWatsonxAssistant size={20} />
-                  </HeaderGlobalAction>
-                  <HeaderGlobalAction
-                    aria-label="Help"
-                    tooltipHighContrast={false}>
-                    <Help size={20} />
-                  </HeaderGlobalAction>
-                  <HeaderGlobalAction
-                    aria-label="Notifications"
-                    tooltipHighContrast={false}>
-                    <Notification size={20} />
-                  </HeaderGlobalAction>
-                  <HeaderGlobalAction
-                    aria-label="Profile"
-                    tooltipHighContrast={false}>
-                    <UserAvatar size={20} />
+                    aria-label={
+                      isSwitcherExpanded ? 'Close switcher' : 'Open switcher'
+                    }
+                    aria-expanded={isSwitcherExpanded}
+                    isActive={isSwitcherExpanded}
+                    onClick={onClickSwitcherExpand}
+                    tooltipAlignment="end"
+                    tooltipHighContrast={false}
+                    id="switcher-button">
+                    <SwitcherIcon size={20} />
                   </HeaderGlobalAction>
                 </HeaderGlobalBar>
+                <HeaderPanel
+                  expanded={isSwitcherExpanded}
+                  onHeaderPanelFocus={onClickSwitcherExpand}
+                  href="#switcher-button">
+                  <Switcher
+                    aria-label="Switcher Container"
+                    expanded={isSwitcherExpanded}>
+                    <SwitcherItem aria-label="Link 1" href="#">
+                      Link 1
+                    </SwitcherItem>
+                    <SwitcherDivider />
+                    <SwitcherItem href="#" aria-label="Link 2">
+                      Link 2
+                    </SwitcherItem>
+                    <SwitcherItem href="#" aria-label="Link 3">
+                      Link 3
+                    </SwitcherItem>
+                    <SwitcherItem href="#" aria-label="Link 4">
+                      Link 4
+                    </SwitcherItem>
+                    <SwitcherItem href="#" aria-label="Link 5">
+                      Link 5
+                    </SwitcherItem>
+                    <SwitcherDivider />
+                    <SwitcherItem href="#" aria-label="Link 6">
+                      Link 6
+                    </SwitcherItem>
+                  </Switcher>
+                </HeaderPanel>
               </Header>
               <SideNav
                 aria-label="Side navigation1"