+++ packages/react/src/components/AnimatedHeader/components/AnimatedHeader/AnimatedHeader.tsx
@@ -7,7 +7,7 @@
  * LICENSE file in the root directory of this source tree.
  */
 import PropTypes from 'prop-types';
-import React, { useEffect, useRef, useState } from 'react';
+import React, { ReactNode, useEffect, useRef, useState } from 'react';
 import {
   Grid,
   Column,
@@ -43,12 +43,13 @@ export interface SelectedWorkspace {
 }
 
 export interface Tile {
-  href: string | null;
+  href?: string | null;
   id: string;
-  mainIcon: string | null;
+  mainIcon?: string | null;
   secondaryIcon?: string | null;
   subtitle?: string | null;
-  title: string | null;
+  title?: string | null;
+  customContent?: ReactNode | null;
 }
 
 export interface TileGroup {
@@ -61,8 +62,10 @@ export interface AnimatedHeaderProps {
   allTiles: TileGroup[];
   allWorkspaces?: SelectedWorkspace[];
   description?: string;
-  handleHeaderItemsToString: (item: TileGroup | null) => string;
-  handleWorkspaceItemsToString: (item: SelectedWorkspace | null) => string;
+  handleHeaderItemsToString?: (item: TileGroup | null) => string;
+  renderHeaderSelectedItem?: (item: TileGroup | null) => ReactNode;
+  handleWorkspaceItemsToString?: (item: SelectedWorkspace | null) => string;
+  renderWorkspaceSelectedItem?: (item: SelectedWorkspace | null) => ReactNode;
   headerAnimation?: object;
   headerStatic?: React.JSX.Element;
   productName?: string;
@@ -74,6 +77,8 @@ export interface AnimatedHeaderProps {
   userName?: string;
   welcomeText?: string;
   workspaceLabel?: string;
+  expandButtonLabel?: string;
+  collapseButtonLabel?: string;
 }
 
 const AnimatedHeader: React.FC<AnimatedHeaderProps> = ({
@@ -85,6 +90,8 @@ const AnimatedHeader: React.FC<AnimatedHeaderProps> = ({
   headerAnimation,
   headerStatic,
   productName = '[Product name]',
+  renderHeaderSelectedItem,
+  renderWorkspaceSelectedItem,
   selectedTileGroup,
   selectedWorkspace,
   setSelectedTileGroup,
@@ -93,6 +100,8 @@ const AnimatedHeader: React.FC<AnimatedHeaderProps> = ({
   userName,
   welcomeText,
   workspaceLabel = `Open in: ${userName}'s workspace` || `Select a workspace`,
+  expandButtonLabel = 'Expand',
+  collapseButtonLabel = 'Collapse',
 }: AnimatedHeaderProps) => {
   const prefix = usePrefix();
   const blockClass = `${prefix}--animated-header`;
@@ -180,8 +189,8 @@ const AnimatedHeader: React.FC<AnimatedHeaderProps> = ({
         <div className={`${blockClass}__lottie-animation--container`}>
           <div
             ref={animationContainer}
-            className={`${blockClass}__lottie-animation ${
-              !open && lottieCollapsed
+            className={`${blockClass}__lottie-animation${
+              !open ? ` ${lottieCollapsed}` : ''
             }`}></div>
         </div>
 
@@ -207,13 +216,11 @@ const AnimatedHeader: React.FC<AnimatedHeaderProps> = ({
             sm={4}
             md={8}
             lg={4}
-            className={`${blockClass}__left-area-container ${
-              !open && descriptionCollapsed
+            className={`${blockClass}__left-area-container${
+              !open ? ` ${descriptionCollapsed}` : ''
             }`}>
             {description && (
-              <Tooltip align="bottom" label={description}>
-                <h2 className={`${blockClass}__description`}>{description}</h2>
-              </Tooltip>
+              <h2 className={`${blockClass}__description`}>{description}</h2>
             )}
 
             {tasksConfig?.button?.text && (
@@ -240,8 +247,13 @@ const AnimatedHeader: React.FC<AnimatedHeaderProps> = ({
                     hideLabel
                     type="inline"
                     items={allTiles}
-                    itemToString={(item) => handleHeaderItemsToString(item)}
                     onChange={(e) => setSelectedTileGroup(e)}
+                    {...(handleHeaderItemsToString
+                      ? { itemToString: handleHeaderItemsToString }
+                      : {})}
+                    {...(renderHeaderSelectedItem
+                      ? { renderSelectedItem: renderHeaderSelectedItem }
+                      : {})}
                   />
                 </div>
               )}
@@ -252,8 +264,8 @@ const AnimatedHeader: React.FC<AnimatedHeaderProps> = ({
           <Column sm={4} md={8} lg={12} className={`${blockClass}__content`}>
             {allWorkspaces && (
               <div
-                className={`${blockClass}__workspace--container ${
-                  !open && contentCollapsed
+                className={`${blockClass}__workspace--container${
+                  !open ? ` ${contentCollapsed}` : ''
                 }`}>
                 <Dropdown
                   id={`${blockClass}__workspace`}
@@ -264,8 +276,16 @@ const AnimatedHeader: React.FC<AnimatedHeaderProps> = ({
                   hideLabel
                   type="inline"
                   items={allWorkspaces}
-                  itemToString={(item) => handleWorkspaceItemsToString(item)}
                   onChange={(e) => setSelectedWorkspace(e)}
+                  {...(handleWorkspaceItemsToString
+                    ? { itemToString: handleWorkspaceItemsToString }
+                    : {})}
+                  {...(renderWorkspaceSelectedItem
+                    ? { renderSelectedItem: renderWorkspaceSelectedItem }
+                    : {})}
+                  {...(selectedWorkspace
+                    ? { selectedItem: selectedWorkspace }
+                    : {})}
                 />
               </div>
             )}
@@ -281,7 +301,8 @@ const AnimatedHeader: React.FC<AnimatedHeaderProps> = ({
                     secondaryIcon={tile.secondaryIcon}
                     title={tile.title}
                     subtitle={tile.subtitle}
-                    productName={productName}></BaseTile>
+                    productName={productName}
+                    customContent={tile.customContent}></BaseTile>
                 );
               })}
             </div>
@@ -296,7 +317,7 @@ const AnimatedHeader: React.FC<AnimatedHeaderProps> = ({
             kind="ghost"
             renderIcon={open ? ChevronUp : ChevronDown}
             onClick={handleButtonCollapseClick}>
-            {open ? `Collapse` : `Expand`}
+            {open ? collapseButtonLabel : expandButtonLabel}
           </Button>
         </div>
       </Grid>
@@ -322,9 +343,19 @@ const AnimatedHeader: React.FC<AnimatedHeaderProps> = ({
   className: PropTypes.string,
 
   /**
+   * Custom collapse button label
+   */
+  collapseButtonLabel: PropTypes.string,
+
+  /**
    * Provide short sentence in max. 3 lines related to product context
    */
   description: PropTypes.string,
+
+  /**
+   * Custom expand button label
+   */
+  expandButtonLabel: PropTypes.string,
 
   /**
    * Helper function passed to downshift that allows the library to render a
@@ -358,6 +389,20 @@ const AnimatedHeader: React.FC<AnimatedHeaderProps> = ({
    * Provide current product name
    */
   productName: PropTypes.string,
+
+  /**
+   * Helper function passed to downshift that allows the library to render a
+   * selected item to as an arbitrary ReactNode. By default it uses standard Carbon renderer that renders only item.label text
+   * (Dropdown under description in header)
+   */
+  renderHeaderSelectedItem: PropTypes.func,
+
+  /**
+   * Helper function passed to downshift that allows the library to render a
+   * selected item to as an arbitrary ReactNode. By default it uses standard Carbon renderer that renders only item.label text
+   * (Dropdown related to workspace selection)
+   */
+  renderWorkspaceSelectedItem: PropTypes.func,
 
   /**
    * The tile group that is active in the header