+++ packages/react/CHANGELOG.md
@@ -3,6 +3,134 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.59.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react@0.58.0...@carbon-labs/react@0.59.0) (2025-05-08)
+
+
+### Bug Fixes
+
+* correct package exports ([#539](https://github.com/carbon-design-system/carbon-labs/issues/539)) ([ef7557e](https://github.com/carbon-design-system/carbon-labs/commit/ef7557e10c1385a6bdc60d32361ce7ba2dad263c))
+
+
+
+
+
+# [0.58.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react@0.57.0...@carbon-labs/react@0.58.0) (2025-05-06)
+
+
+### Features
+
+* **chat:** extensive updates to code charts and chat ([#582](https://github.com/carbon-design-system/carbon-labs/issues/582)) ([73772da](https://github.com/carbon-design-system/carbon-labs/commit/73772da806e67eba7314f0de1c17cb0241f3529f))
+
+
+
+
+
+# [0.57.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react@0.56.0...@carbon-labs/react@0.57.0) (2025-04-29)
+
+**Note:** Version bump only for package @carbon-labs/react
+
+
+
+
+
+# [0.56.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react@0.55.0...@carbon-labs/react@0.56.0) (2025-04-25)
+
+**Note:** Version bump only for package @carbon-labs/react
+
+
+
+
+
+# [0.55.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react@0.54.0...@carbon-labs/react@0.55.0) (2025-04-22)
+
+
+### Features
+
+* create split panel component ([#501](https://github.com/carbon-design-system/carbon-labs/issues/501)) ([a579b97](https://github.com/carbon-design-system/carbon-labs/commit/a579b978a085b38f4d2f7ea4b853a29e022632b0))
+
+
+
+
+
+# [0.54.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react@0.53.0...@carbon-labs/react@0.54.0) (2025-04-22)
+
+
+### Features
+
+* **UIShell:** add TrialCountdown component ([#545](https://github.com/carbon-design-system/carbon-labs/issues/545)) ([b5e4f33](https://github.com/carbon-design-system/carbon-labs/commit/b5e4f33eef2eaac609bcf5e720c6477c6032b607))
+
+
+
+
+
+# [0.53.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react@0.52.0...@carbon-labs/react@0.53.0) (2025-04-18)
+
+**Note:** Version bump only for package @carbon-labs/react
+
+
+
+
+
+# [0.52.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react@0.51.0...@carbon-labs/react@0.52.0) (2025-04-10)
+
+
+### Bug Fixes
+
+* **animated-header:** add backup for autoprefixer ([#563](https://github.com/carbon-design-system/carbon-labs/issues/563)) ([8c3c4a6](https://github.com/carbon-design-system/carbon-labs/commit/8c3c4a6846627b6fbb8e65e50bb731acb52784dc))
+
+
+
+
+
+# [0.51.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react@0.50.0...@carbon-labs/react@0.51.0) (2025-04-10)
+
+
+### Bug Fixes
+
+* **animated-header:** cleanup Safari issues ([#556](https://github.com/carbon-design-system/carbon-labs/issues/556)) ([530e01a](https://github.com/carbon-design-system/carbon-labs/commit/530e01ae13d94391732bf34a3b26edb87b9ea161))
+
+
+
+
+
+# [0.50.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react@0.49.0...@carbon-labs/react@0.50.0) (2025-04-08)
+
+
+### Features
+
+* **animated-header:** remove strictmode from example app + tooltip ([#548](https://github.com/carbon-design-system/carbon-labs/issues/548)) ([c969903](https://github.com/carbon-design-system/carbon-labs/commit/c969903eab1bd7b78fbb3b7f9770e4839b1d7f6b))
+
+
+
+
+
+# [0.49.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react@0.48.0...@carbon-labs/react@0.49.0) (2025-04-07)
+
+
+### Bug Fixes
+
+* **UIShell:** remove box-shadow on header action focus ([#554](https://github.com/carbon-design-system/carbon-labs/issues/554)) ([55a8671](https://github.com/carbon-design-system/carbon-labs/commit/55a8671faf1116f33b27260bef0f5dba2802aed5))
+
+
+
+
+
+# [0.48.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react@0.47.0...@carbon-labs/react@0.48.0) (2025-04-07)
+
+**Note:** Version bump only for package @carbon-labs/react
+
+
+
+
+
+# [0.47.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react@0.46.0...@carbon-labs/react@0.47.0) (2025-04-04)
+
+**Note:** Version bump only for package @carbon-labs/react
+
+
+
+
+
 # [0.46.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react@0.45.0...@carbon-labs/react@0.46.0) (2025-04-02)