+++ packages/react/src/components/ThemeSettings/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.5.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react-theme-settings@0.4.0...@carbon-labs/react-theme-settings@0.5.0) (2025-05-08)
+
+
+### Bug Fixes
+
+* correct package exports ([#539](https://github.com/carbon-design-system/carbon-labs/issues/539)) ([ef7557e](https://github.com/carbon-design-system/carbon-labs/commit/ef7557e10c1385a6bdc60d32361ce7ba2dad263c))
+
+
+
+
+
+# [0.4.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react-theme-settings@0.3.0...@carbon-labs/react-theme-settings@0.4.0) (2025-04-04)
+
+**Note:** Version bump only for package @carbon-labs/react-theme-settings
+
+
+
+
+
 # [0.3.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/react-theme-settings@0.2.0...@carbon-labs/react-theme-settings@0.3.0) (2025-03-26)