+++ packages/web-components/src/components/chat/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.18.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/ai-chat@0.17.0...@carbon-labs/ai-chat@0.18.0) (2025-05-08)
+
+
+### Bug Fixes
+
+* correct package exports ([#539](https://github.com/carbon-design-system/carbon-labs/issues/539)) ([ef7557e](https://github.com/carbon-design-system/carbon-labs/commit/ef7557e10c1385a6bdc60d32361ce7ba2dad263c))
+
+
+
+
+
+# [0.17.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/ai-chat@0.16.0...@carbon-labs/ai-chat@0.17.0) (2025-05-06)
+
+
+### Features
+
+* **chat:** extensive updates to code charts and chat ([#582](https://github.com/carbon-design-system/carbon-labs/issues/582)) ([73772da](https://github.com/carbon-design-system/carbon-labs/commit/73772da806e67eba7314f0de1c17cb0241f3529f))
+
+
+
+
+
+# [0.16.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/ai-chat@0.15.0...@carbon-labs/ai-chat@0.16.0) (2025-04-04)
+
+**Note:** Version bump only for package @carbon-labs/ai-chat
+
+
+
+
+
 # [0.15.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/ai-chat@0.13.0-rc.1...@carbon-labs/ai-chat@0.15.0) (2025-01-30)