+++ packages/web-components/src/components/network-graph/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.14.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/network-graph@0.13.0...@carbon-labs/network-graph@0.14.0) (2025-05-08)
+
+
+### Bug Fixes
+
+* correct package exports ([#539](https://github.com/carbon-design-system/carbon-labs/issues/539)) ([ef7557e](https://github.com/carbon-design-system/carbon-labs/commit/ef7557e10c1385a6bdc60d32361ce7ba2dad263c))
+
+
+
+
+
+# [0.13.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/network-graph@0.12.0...@carbon-labs/network-graph@0.13.0) (2025-04-04)
+
+**Note:** Version bump only for package @carbon-labs/network-graph
+
+
+
+
+
 # [0.12.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/network-graph@0.10.0-rc.0...@carbon-labs/network-graph@0.12.0) (2025-01-30)